Fix bug display incorrect screen when connection is lost
