+++ packages/extension-koni-ui/src/Popup/Home/ChainBalances/ChainBalanceDetail/ChainBalanceDetail.tsx
@@ -1,9 +1,12 @@
 // Copyright 2019-2022 @polkadot/extension-koni-ui authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
+import { faArrowLeft } from '@fortawesome/free-solid-svg-icons';
+import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import React, { useCallback, useEffect, useRef, useState } from 'react';
 import styled from 'styled-components';
 
+import useTranslation from '@polkadot/extension-koni-ui/hooks/useTranslation';
 import { ThemeProps } from '@polkadot/extension-koni-ui/types';
 import { AccountInfoByNetwork, BalanceInfo } from '@polkadot/extension-koni-ui/util/types';
 
@@ -13,6 +16,7 @@ import ChainBalanceDetailItem from '../ChainBalanceDetail/ChainBalanceDetailItem
 interface Props extends ThemeProps {
   accountInfo: AccountInfoByNetwork;
   balanceInfo: BalanceInfo;
+  backToHome: () => void;
   className?: string;
   setQrModalOpen: (visible: boolean) => void;
   setQrModalProps: (props: {
@@ -23,9 +27,10 @@ interface Props extends ThemeProps {
   }) => void;
 }
 
-function ChainBalanceDetail ({ accountInfo, balanceInfo, className, setQrModalOpen, setQrModalProps }: Props): React.ReactElement<Props> {
+function ChainBalanceDetail ({ accountInfo, backToHome, balanceInfo, className, setQrModalOpen, setQrModalProps }: Props): React.ReactElement<Props> {
   const [selectedNetworkKey, setSelectedNetworkKey] = useState<string>('');
   const ref = useRef<HTMLDivElement>(null);
+  const { t } = useTranslation();
 
   useEffect(() => {
     ref.current?.scrollIntoView({ behavior: 'smooth', block: 'start' });
@@ -44,6 +49,19 @@ function ChainBalanceDetail ({ accountInfo, balanceInfo, className, setQrModalOp
       className={className}
       ref={ref}
     >
+      {accountInfo && balanceInfo &&
+      <div
+        className='chain-balance-detail__back-btn'
+        onClick={backToHome}
+      >
+        <FontAwesomeIcon
+          className='chain-balance-detail__back-icon'
+          // @ts-ignore
+          icon={faArrowLeft}
+        />
+        <span>{t<string>('Back to home')}</span>
+      </div>
+      }
       <ChainBalanceDetailItem
         accountInfo={accountInfo}
         balanceInfo={balanceInfo}
@@ -90,5 +108,19 @@ export default React.memo(styled(ChainBalanceDetail)(({ theme }: Props) => `
       display: block;
       background: ${theme.boxBorderColor};
     }
+  }
+
+  .chain-balance-detail__back-btn {
+    color: ${theme.buttonTextColor2};
+    font-size: 15px;
+    line-height: 26px;
+    font-weight: 500;
+    margin-left: 25px;
+    cursor: pointer;
+    margin-bottom: 10px;
+  }
+
+  .chain-balance-detail__back-icon {
+    padding-right: 7px;
   }
 `));
