+++ packages/extension-koni-ui/src/Popup/Home/ChainBalances/ChainBalances.tsx
@@ -111,6 +111,10 @@ function ChainBalances ({ address,
     setShowBalanceDetail(true);
   }, [setShowBalanceDetail]);
 
+  const _backToHome = useCallback(() => {
+    setShowBalanceDetail(false);
+  }, [setShowBalanceDetail]);
+
   const toggleBalanceDetail = useCallback((networkKey: string) => {
     if (networkKey === selectedNetworkKey) {
       setSelectedNetworkKey('');
@@ -230,6 +234,7 @@ function ChainBalances ({ address,
           <>
             <ChainBalanceDetail
               accountInfo={selectedInfo}
+              backToHome={_backToHome}
               balanceInfo={selectedBalanceInfo}
               setQrModalOpen={setQrModalOpen}
               setQrModalProps={setQrModalProps}
@@ -237,7 +242,6 @@ function ChainBalances ({ address,
           </>
         )
       }
-
     </div>
   );
 }
